stan-backend/requirements.txt
# Modern requirements with latest AI models
# Core AI - Updated to latest Gemini
google-generativeai>=0.8.3

# Web Framework with streaming support
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sse-starlette>=2.1.0

# Database
supabase>=2.7.0

# Environment
python-dotenv>=1.0.1

# Essential dependencies
pydantic>=2.9.0
httpx>=0.28.1
aiohttp>=3.10.5

# Caching (REQUIRED for production - cost optimization)
redis>=5.0.0

# Structured Logging
structlog>=24.1.0
python-json-logger>=2.0.7

# Rate Limiting
slowapi>=0.1.9

# Error Tracking (optional but recommended)
sentry-sdk[fastapi]>=1.40.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.28.1  # For testing async endpoints
